Refactor de Storages a una clase base
