+++ infra/apps/catalog/storage/distribution_storage.py
@@ -1,7 +1,7 @@
 import os
 
 from infra.apps.catalog.constants import CATALOG_ROOT
-from infra.apps.catalog.storage.catalog_storage import CustomCatalogStorage
+from infra.apps.catalog.storage.infra_storage import InfraStorage
 
 
 def distribution_directory(instance):
@@ -14,8 +14,8 @@ def distribution_directory(instance):
                         'download')
 
 
-class DistributionStorage(CustomCatalogStorage):
-    def _latest_file_path(self, instance):
+class DistributionStorage(InfraStorage):
+    def latest_file_path(self, instance):
         name = instance.name.split('/')[-1]
         return os.path.join(distribution_directory(instance),
                             name)
